Removed commented out methods and improved code structure

+++ Classes/Canvas/CanvasDataSourceProtocol.h
@@ -1,17 +1,11 @@
-//
-//  CanvasDataSourceProtocol.h
-//  CanvasKit
-//
 //  Created by Jason Morrissey
-
 
 #import "CanvasTileView.h"
 
 @protocol CanvasDataSourceProtocol
 
 - (CanvasTileView *) tileViewForIndex:(long) tileIndex;
-//- (NSMutableArray *) tileDictionariesInRange:(NSRange) range;
-- (long) totalNumberOfTiles;
+- (NSUInteger) totalNumberOfTiles;
 - (CGSize) tileDimensions;
 - (CGSize) pageMargin;
